@@ -11,19 +11,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "^12.2.14",
-    "@angular/animations": "~12.2.15",
-    "@angular/common": "~12.2.15",
-    "@angular/compiler": "~12.2.15",
-    "@angular/core": "~12.2.15",
-    "@angular/forms": "~12.2.15",
-    "@angular/platform-browser": "~12.2.15",
-    "@angular/platform-browser-dynamic": "~12.2.15",
-    "@angular/router": "~12.2.15",
+    "@angular-devkit/schematics": "^13.3.11",
+    "@angular/animations": "~13.4.0",
+    "@angular/common": "~13.4.0",
+    "@angular/compiler": "~13.4.0",
+    "@angular/core": "~13.4.0",
+    "@angular/forms": "~13.4.0",
+    "@angular/localize": "~13.4.0",
+    "@angular/platform-browser": "~13.4.0",
+    "@angular/platform-browser-dynamic": "~13.4.0",
+    "@angular/router": "~13.4.0",
+    "@ng-bootstrap/ng-bootstrap": "^12.1.2",
     "@ngrx/effects": "^13.0.2",
     "@ngrx/schematics": "^13.0.2",
     "@ngrx/store": "^13.0.2",
     "@ngrx/store-devtools": "^13.0.2",
+    "@popperjs/core": "^2.10.2",
     "bootstrap": "^4.5.2",
     "ngrx": "0.0.1-security",
     "rxjs": "~6.6.7",
@@ -31,15 +34,15 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.14",
-    "@angular/cli": "~12.2.14",
-    "@angular/compiler-cli": "~12.2.15",
-    "@angular/language-service": "~12.2.15",
+    "@angular-devkit/build-angular": "~13.3.11",
+    "@angular/cli": "~13.3.11",
+    "@angular/compiler-cli": "~13.4.0",
+    "@angular/language-service": "~13.4.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
+    "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~6.3.9",
     "karma-chrome-launcher": "~3.1.0",
@@ -49,6 +52,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.6.4"
   }
 }
