@@ -1,10 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HomeComponent } from './components/home/home.component';
+import { PeopleNgrxComponent } from './components/people-ngrx/people-ngrx.component'
 
 const routes: Routes = [
-  {path: '', redirectTo: '/home', pathMatch: 'full'},
-  { path: 'home', component: HomeComponent},
+  { path: '', redirectTo: '/home', pathMatch: 'full' },
+  { path: 'people-ngrx', component: PeopleNgrxComponent },
+  { path: 'home', component: HomeComponent },
 ];
 
 @NgModule({
